+++ ext/vesper-cli/src/edu/ucsc/refactor/cli/commands/OriginShow.java
@@ -1,0 +1,28 @@
+package edu.ucsc.refactor.cli.commands;
+
+import com.google.common.base.Objects;
+import edu.ucsc.refactor.cli.Environment;
+import edu.ucsc.refactor.cli.Result;
+import edu.ucsc.refactor.cli.VesperCommand;
+import io.airlift.airline.Command;
+
+/**
+ * @author hsanchez@cs.ucsc.edu (Huascar A. Sanchez)
+ */
+@Command(name = "show", description = "Shows the current indexed source")
+public class OriginShow extends VesperCommand {
+    @Override public Result execute(Environment environment) throws RuntimeException {
+        ensureValidState(environment);
+
+        if(!environment.containsOrigin()){
+            return Result.nothing();
+        }
+
+        return Result.sourcePackage(environment.getOrigin());
+    }
+
+    @Override public String toString() {
+        return Objects.toStringHelper("OriginShow")
+                .toString();
+    }
+}
