+++ ext/vesper-cli/src/edu/ucsc/refactor/cli/commands/RemoveClassCommand.java
@@ -1,0 +1,15 @@
+package edu.ucsc.refactor.cli.commands;
+
+import edu.ucsc.refactor.ChangeRequest;
+import edu.ucsc.refactor.SourceSelection;
+import io.airlift.airline.Command;
+
+/**
+ * @author hsanchez@cs.ucsc.edu (Huascar A. Sanchez)
+ */
+@Command(name = "class", description = "Remove unused inner class from the indexed Source")
+public class RemoveClassCommand extends RemoveMemberCommand {
+    @Override protected ChangeRequest createChangeRequest(SourceSelection selection) {
+        return ChangeRequest.deleteClass(selection);
+    }
+}
