+++ ext/vesper-cli/src/edu/ucsc/refactor/cli/commands/RemoveMethodCommand.java
@@ -1,0 +1,15 @@
+package edu.ucsc.refactor.cli.commands;
+
+import edu.ucsc.refactor.ChangeRequest;
+import edu.ucsc.refactor.SourceSelection;
+import io.airlift.airline.Command;
+
+/**
+ * @author hsanchez@cs.ucsc.edu (Huascar A. Sanchez)
+ */
+@Command(name = "method", description = "Remove unused method from the indexed Source")
+public class RemoveMethodCommand extends RemoveMemberCommand {
+    @Override protected ChangeRequest createChangeRequest(SourceSelection selection) {
+        return ChangeRequest.deleteMethod(selection);
+    }
+}
