+++ ext/vesper-cli/src/edu/ucsc/refactor/cli/commands/RemoveParameterCommand.java
@@ -1,0 +1,15 @@
+package edu.ucsc.refactor.cli.commands;
+
+import edu.ucsc.refactor.ChangeRequest;
+import edu.ucsc.refactor.SourceSelection;
+import io.airlift.airline.Command;
+
+/**
+ * @author hsanchez@cs.ucsc.edu (Huascar A. Sanchez)
+ */
+@Command(name = "param", description = "Remove unused param from a method in the indexed Source")
+public class RemoveParameterCommand extends RemoveMemberCommand {
+    @Override protected ChangeRequest createChangeRequest(SourceSelection selection) {
+        return ChangeRequest.deleteParameter(selection);
+    }
+}
